UnstCInfo구조체의 fontSize가 int형이 아닌 float형이었던 오류를 수정한다
int형이 올바른 자료형이기 때문에, 수정한다.

+++ Assets/UnstableObject/Scripts/UnstableText.cs
@@ -16,11 +16,11 @@ public struct UnstCInfo
 
     public Font font;
 
-    public float fontSize;
+    public int   fontSize;
     public float rotation;
     public float vibration;
 
-    public UnstCInfo(Color color, FontStyle fontStyle, Font font, UnstableObject unstableObject, float fontSize)
+    public UnstCInfo(Color color, FontStyle fontStyle, Font font, UnstableObject unstableObject, int fontSize)
     {
         this.color     = color;
         this.fontStyle = fontStyle;
